+++ src/planuml.ts
@@ -8,6 +8,7 @@ import { previewer } from './previewer';
 import { builder } from "./builder";
 import { symboler } from "./symboler";
 import { urlMaker } from "./urlMaker";
+import { formatter } from "./format/formatter";
 
 export var outputPanel = vscode.window.createOutputChannel("PlantUML");
 export var context: vscode.ExtensionContext;
@@ -34,6 +35,8 @@ export class PlantUML {
             ds.push(...symboler.register());
             //register server
             ds.push(...urlMaker.register());
+            //register formatter
+            ds.push(...formatter.register());
             return ds;
         } catch (error) {
             outputPanel.clear()